requirements.txt
psycopg2-binary==2.9.9
Django>=4.2
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
invoke==2.2.0
redis==5.0.1
celery==5.3.5
django-redis==5.4.0
